requirements.txt
# pip install aws-cdk
aws-cdk-lib==2.10.0

# pip install elasticsearch
elasticsearch>=7.0.0,<7.11
requests-aws4auth==0.9

# pip install gremlinpython
aenum==2.2.2
gremlinpython==3.4.4
isodate==0.6.0
tornado==4.5.3

# pip install redis
redis==3.3.11

# misc
pretty-errors==1.2.19
